refactor(jina): 음악 선호 선택 모달 features로 이동
