+++ src/features/musicPreference/model/type.ts
