+++ src/pages/demo/index.tsx
@@ -2,7 +2,9 @@ import { useState } from 'react';
 
 import { MainLayout } from 'widgets/ui/';
 
-import { Main, MusicSelectionModal } from 'shared/ui/';
+import { MusicSelectionModal } from 'features/musicPreference/ui/MusicPreferenceModal';
+
+import { Main } from 'shared/ui/';
 
 // 컴포넌트 렌더링 테스트를 위한 /demo 페이지
 const Demo = () => {
